Ejercicios resueltos hasta el 4.21

+++ controllers/blogs.js
@@ -17,8 +17,55 @@ blogsRouter.get("/:id", async (request, response) => {
   }
 });
 
+blogsRouter.post("/", async (request, response) => {
+  const body = request.body;
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "invalid token" });
+  }
+  const user = await User.findById(decodedToken.id);
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes || 0,
+    user: user._id,
+  });
+
+  const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
+  response.status(201).json(savedBlog);
+});
+
 blogsRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "invalid token" });
+  }
+  const userRequestingDelete = await User.findById(decodedToken.id);
+  const blogToDelete = await Blog.findById(request.params.id);
+
+  if (!blogToDelete) {
+    return response.status(404).json({ error: "blog not found" });
+  }
+
+  const creatorBlogId = blogToDelete.user.toString();
+
+  if (!(userRequestingDelete._id.toString() === creatorBlogId)) {
+    return response
+      .status(401)
+      .json({ error: "you don't have permission to delete this blog" });
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
+  userRequestingDelete.blogs = userRequestingDelete.blogs.filter(
+    (blogId) => blogId.toString() !== request.params.id
+  );
+  await userRequestingDelete.save();
+
   response.status(204).end();
 });
 
@@ -37,37 +84,6 @@ blogsRouter.put("/:id", async (request, response) => {
     runValidators: true,
   });
   response.json(updatedBlog);
-});
-
-const getTokenFrom = (request) => {
-  const authorization = request.get("authorization");
-  if (authorization && authorization.startsWith("Bearer ")) {
-    return authorization.replace("Bearer ", "");
-  }
-  return null;
-};
-
-blogsRouter.post("/", async (request, response) => {
-  const body = request.body;
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({error: 'invalid token'})
-  }  
-  const user = await User.findById(decodedToken.id);
-
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0,
-    user: user._id,
-  });
-
-  const savedBlog = await blog.save();
-  user.blogs = user.blogs.concat(savedBlog._id);
-  await user.save();
-
-  response.status(201).json(savedBlog);
 });
 
 module.exports = blogsRouter;
